Refactor out waitForBackend
